[NAYB-84] refactor : CategoryFixture 경로 변경
